Fix isochrone parameter date_time conversion to int, not float.
According to the isochrone API reference value of
date_time::type can be 0, 1, or 2.

+++ src/worker.cc
@@ -468,7 +468,7 @@ namespace {
     parse_locations(doc, options.mutable_avoid_locations(), "avoid_locations", 133, track);
 
     //time type
-    auto date_time_type = rapidjson::get_optional<float>(doc, "/date_time/type");
+    auto date_time_type = rapidjson::get_optional<int>(doc, "/date_time/type");
     if(date_time_type) {
       options.set_date_time_type(static_cast<valhalla::odin::DirectionsOptions::DateTimeType>(*date_time_type));
     }
